feat: Split syscall APIs with core implementations (#19)
This change defines a core set of CKB syscall APIs, it then separates
utility functions that can be built on top of the core syscall
APIs. As a result, the implementation of the core syscall APIs can be
swapped, without affecting the utility functions one layer above. This
enables simulator based solution, where we swap the core syscall APIs
with dummy functions, so the same smart contracts can be built
differently on multiple platforms: a RISC-V based version for
execution on chain, and a x86 version for profiling and auditing reason.